`Uint16` replaced by `Uint8`

+++ js/main.js
@@ -166,7 +166,7 @@ function logToTerminal(message) {
 
 function str2ab(str) {
   let buf = new ArrayBuffer(str.length * 2); // 2 bytes for each char
-  let bufView = new Uint16Array(buf);
+  let bufView = new Uint8Array(buf);
 
   for (let i = 0, strLen = str.length; i < strLen; i++) {
     bufView[i] = str.charCodeAt(i);
